+++ ModulesPath/sessobj/decoders.py
@@ -10,71 +10,12 @@ from scipy.ndimage import gaussian_filter, gaussian_filter1d
 from scipy.special import factorial
 from tqdm import tqdm
 
-from .parsePath import Recinfo
-from .pfPlot import pf1d, pf2d
-from .plotUtil import Fig
+from ..parsePath import Recinfo
+from ..placefields import PF1d, PF2d
+from ..plotUtil import Fig
 
 
-class Bayes:
-    binsize = 0.02
-    n_jobs = 8
-
-    def __init__(self):
-        self._events = None
-        self.posterior = None
-        self.decodedPos = None
-
-    @property
-    def events(self):
-        return self._events
-
-    @events.setter
-    def events(self, events: pd.DataFrame):
-
-        if isinstance(events, pd.DataFrame):
-            assert (
-                pd.Series(["start", "end"]).isin(events.columns).all()
-            ), "events should have start and end as column names"
-        elif isinstance(events, np.ndarray):
-            events = pd.DataFrame({"start": events[:, 0], "end": events[:, 1]})
-        self._events = events
-
-    def _decoder(self, spkcount, ratemaps):
-        """
-        ===========================
-        Probability is calculated using this formula
-        prob = (1 / nspike!)* ((tau * frate)^nspike) * exp(-tau * frate)
-        where,
-            tau = binsize
-        ===========================
-        """
-        tau = self.binsize
-        nCells = spkcount.shape[0]
-        cell_prob = np.zeros((ratemaps.shape[1], spkcount.shape[1], nCells))
-        for cell in range(nCells):
-            cell_spkcnt = spkcount[cell, :][np.newaxis, :]
-            cell_ratemap = ratemaps[cell, :][:, np.newaxis]
-
-            coeff = 1 / (factorial(cell_spkcnt))
-            # broadcasting
-            cell_prob[:, :, cell] = (((tau * cell_ratemap) ** cell_spkcnt) * coeff) * (
-                np.exp(-tau * cell_ratemap)
-            )
-
-        posterior = np.prod(cell_prob, axis=2)
-        posterior /= np.sum(posterior, axis=0)
-
-        return posterior
-
-
-class DecodeBehav:
-    def __init__(self, pf1d_obj: pf1d, pf2d_obj: pf2d):
-
-        self.bayes1d = Bayes1d(pf1d_obj)
-        self.bayes2d = bayes2d(pf2d_obj)
-
-
-class Bayes1d:
+class Decode1d:
     binsize = 0.02
     n_jobs = 8
 
@@ -466,7 +407,7 @@ class Bayes1d:
                 axdec.set_xlabel("Time (ms)")
 
 
-class bayes2d(Bayes):
+class Decode2d(Bayes):
     def __init__(self, pf2d_obj: pf2d):
 
         assert isinstance(pf2d_obj, pf2d)
